Fix old e2e tests, refactor and add new e2e tests.
